+++ addons/filamind_mrp_iot/models/__init__.py
@@ -1,2 +1,3 @@
+from . import iot_trigger
 from . import mrp_workcenter
 from . import mrp_workorder