+++ addons/filamind_stock_iot/models/__init__.py
@@ -1,2 +1,3 @@
 from . import stock_warehouse
 from . import stock_picking
+from . import stock_picking_type